+++ .github/workflows/conformance.yml
@@ -149,7 +149,7 @@ jobs:
           required-secrets: AzureKeyVaultName,AzureKeyVaultSecretStoreTenantId,AzureKeyVaultSecretStoreServicePrincipalClientId,AzureKeyVaultSecretStoreServicePrincipalClientSecret
         - component: bindings.azure.cosmosdb
           required-secrets: AzureCosmosDBMasterKey,AzureCosmosDBUrl,AzureCosmosDB,AzureCosmosDBCollection
-        - component: state.cloudflare.kv
+        - component: state.cloudflare.workerskv
         EOF
         )
         echo "cron-components=$CRON_COMPONENTS" >> $GITHUB_OUTPUT
@@ -321,7 +321,7 @@ jobs:
       if: contains(matrix.component, 'rabbitmq')
 
     - name: Setup Cloudflare KV
-      if: matrix.component == 'state.cloudflare.kv'
+      if: matrix.component == 'state.cloudflare.workerskv'
       env:
         CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
         CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
@@ -472,7 +472,7 @@ jobs:
         az sql db delete --resource-group ${{ env.AzureResourceGroupName }} --server ${{ env.AzureSqlServerName }} -n ${{ env.AzureSqlServerDbName }} --yes
 
     - name: Cleanup Cloudflare KV
-      if: always() && matrix.component == 'state.cloudflare.kv'
+      if: always() && matrix.component == 'state.cloudflare.workerskv'
       continue-on-error: true
       env:
         CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
