+++ tests/config/state/tests.yml
@@ -39,7 +39,7 @@ components:
   - component: cassandra
     allOperations: false
     operations: [ "set", "get", "delete", "bulkset", "bulkdelete", "ttl" ]
-  - component: cloudflare.kv
+  - component: cloudflare.workerskv
     allOperations: false
     operations: [ "set", "get", "delete", "bulkset", "bulkdelete"]
   - component: cockroachdb
